% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasbin.R
\name{rasbin_header}
\alias{rasbin_header}
\title{Read the .rasbin File Header}
\usage{
rasbin_header(path)
}
\arguments{
\item{path}{Character. Path to a \code{.rasbin} file.}
}
\value{
Named numeric vector with elements \code{n_samples}, \code{n_snps},
\code{dtype}.
}
\description{
Read the .rasbin File Header
}
