% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/solvemyproblem.R
\name{solvemyproblem}
\alias{solvemyproblem}
\title{Solve my problem}
\usage{
solvemyproblem(x)
}
\arguments{
\item{x}{any problem you have}
}
\value{
the best possible solution will be returned.
}
\description{
You have a problem?
Just use the function solvemyproblem and enter it and you will be presented with the best possilbe solution
}

