\name{RAS-package}
\alias{RAS-package}
\alias{RAS}
\docType{package}
\title{
\packageTitle{RAS}
}
\description{
\packageDescription{RAS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RAS}
\packageIndices{RAS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RAS}

Maintainer: \packageMaintainer{RAS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
